Simplify README
Moved docs out of the README into other docs, such as `using.md` (which
is only just barely a start at usage docs, lots left to do here!)
Hopefully in #22 we could expand on these a bit.

The README which we did have was created specifically for the folks
working on the Build CRD to understand how this project related to the
Build CRD; but if a person comes to this project not already knowing
about the Build CRD it's pretty confusing.

Removed the `status` sections from our run examples since folks wouldn't
add those manually. That was only there b/c initially all we had were
examples, which we had to build the real thing from.

Fixes #136