Add @shashwathi as an Owner! 🎉
@shashwathi has contributed difficult and high quality features to
build-pipeline, including output linking, validation, recording events,
multiple input sources, service accounts, and more. She is one of the
top project contributors, not to mention that she's also one of the top
contributors to knative/build - which she's also an owner of and will
probably be moving into this repo eventually.